backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==1.0.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
pandas==2.1.3